Refactor GUI.py into Modular Components
