+++ constants.py
@@ -1,0 +1,7 @@
+WIDTH = 1500
+HEIGHT = 740
+SETTINGS_FILE = "user_settings.txt"
+SPEAKER_COLORS = {
+    "Speaker 1": "#029CFF",  # Light Blue
+    "Speaker 2": "#FF5733"   # Light Red
+}
